feat: 企业租户核心功能 + RBAC 角色 ID 类型重构
- 实现 createEnterprise 完整流程（企业创建 + OWNER 成员/角色自动写入）
- 实现 listUserEnterprises 列表查询（含角色编码 + 成员数统计）
- 实现 EnterpriseTeamMembers 邀请/列表/修改/移除 CRUD
- EnterpriseStatus 枚举 + @EnumValue 注解完善
- 新增 EnterpriseCreateBO/ListUserEnterprisesBO 业务对象
- 新增 EnterpriseCovert BO→VO 转换器
- Controller 层补充 Swagger 注解

- sys_roles.id 及关联 role_id 字段 BIGINT → INTEGER
- Java 实体/映射器/服务层 Long → Integer 同步
- UserRolesServiceImpl 分配角色防重拦截（代码校验 + 索引兜底）
- enterprises.name 添加唯一部分索引（WHERE is_deleted = false）
- ErrorCode 新增 ROLE_ALREADY_ASSIGNED(20016)

- DDL(Migration): role_id INTEGER → INTEGER, 唯一索引
- 文档: RESTful API 文档同步

+++ modules/system/src/main/java/interview/system/rbac/mapper/PermissionsMapper.java
@@ -19,6 +19,6 @@ import java.util.List;
 @Mapper
 public interface PermissionsMapper extends BaseMapper<SysPermission> {
 
-    List<String> getPermCodeByRoleId(@Param("roleIds") List<Long> roleIds);
+    List<String> getPermCodeByRoleId(@Param("roleIds") List<Integer> roleIds);
 
 }
